WeatherProvider → WXPProvider
